@@ -1,7 +1,7 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 using SharpGo.Game.Players;
 using SharpGo.Game;
-using MathlibNET.Random;
+using DrawingBoardNET.Drawing;
 
 namespace SharpGo.Tests
 {
@@ -13,8 +13,8 @@ namespace SharpGo.Tests
 		{
 			Player p1 = new RandomNoPassPlayer(PlayerColor.Black);
 			Player p2 = new RandomNoPassPlayer(PlayerColor.White);
-			Rng.Seed = 1000000000;
-			GoGame game = new GoGame(p1, p2, boardSize: 19, closeOnGameEnd: true, render: true, frameRate: 999);
+			MathUtils.RandomSeed = 1000000000;
+			GoGame game = new(p1, p2, boardSize: 19, closeOnGameEnd: true, render: true, frameRate: 999);
 			game.Start();
 		}
 	}
