@@ -1,7 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
-using MathlibNET;
-using MathlibNET.Random;
+using DrawingBoardNET.Drawing;
 
 namespace SharpGo.Game.Players
 {
@@ -11,12 +10,12 @@ namespace SharpGo.Game.Players
 
 		protected override (int, int) PickPosition(Board board, HashSet<Intersection> legalIntersections)
 		{
-			int index = Rng.Rand(legalIntersections.Count);
+			int index = MathUtils.Rand(legalIntersections.Count);
 			Intersection intersection = legalIntersections.ToArray()[index];
 			return (intersection.I, intersection.J);
 		}
 
 		protected override bool Pass(Board board, HashSet<Intersection> legalIntersections) =>
-			Rng.Rand() < SpecialFunctions.Lerp(NbTurnsPlayed, 0, 1000, 0, 0.5);
+			MathUtils.Rand() < MathUtils.Lerp(NbTurnsPlayed, 0, 1000, 0, 0.5f);
 	}
 }
